OGM-293 Extracting generic OgmQueryLoader and Query implementation based on BackendQuery
